crowd-count/requirements.txt
torch==2.5.0
gradio==4.44.1
torchvision==0.20.0
numpy==1.26.4
Pillow==10.4.0